minor change to README
I read the previous text which said it was 'primarily used for adding a sidebar' - of course with the new Discourse Core sidebar this use-case is possibly going to become redundant, and from what I understand of the capabilities of Layouts it's significantly underselling it. Just thought I'd update. Happy for alternative proposals.